ff - scz180 sd mode (again)

+++ ff/rc2014/include/newlib/ff.h
@@ -252,7 +252,7 @@ typedef struct {
     WORD    ftime;          /* Modified time */
     BYTE    fattrib;        /* File attribute */
 #if FF_USE_LFN
-    TCHAR    altname[FF_SFN_BUF + 1];  /* Altenative file name */
+    TCHAR    altname[FF_SFN_BUF + 1];  /* Alternative file name */
     TCHAR    fname[FF_LFN_BUF + 1];    /* Primary file name */
 #else
     TCHAR    fname[12 + 1]; /* File name */
@@ -411,10 +411,10 @@ void ff_memfree (void* mblock);            /* Free memory block */
 
 /* Sync functions */
 #if FF_FS_REENTRANT
-int ff_cre_syncobj (BYTE vol, FF_SYNC_t* sobj);    /* Create a sync object */
-int ff_req_grant (FF_SYNC_t sobj);        /* Lock sync object */
+int16_t ff_cre_syncobj (BYTE vol, FF_SYNC_t* sobj);    /* Create a sync object */
+int16_t ff_req_grant (FF_SYNC_t sobj);        /* Lock sync object */
 void ff_rel_grant (FF_SYNC_t sobj);        /* Unlock sync object */
-int ff_del_syncobj (FF_SYNC_t sobj);    /* Delete a sync object */
+int16_t ff_del_syncobj (FF_SYNC_t sobj);    /* Delete a sync object */
 #endif
